Use URL query to fix Chrome not re-fetching image
This solves a bug that occurs in Chrome (as recent as v54).

When the local image is removed from the store, the Gravatar
component shows the avatar URL from
`state.users.items[user_id].avatar_URL`.

This is the same Gravatar URL as before, but the browser cache
is expired so it should be re-downloaded.

Chrome does not re-download the image even though the cache expired,
so we force a re-download by adding a parameter to the URL.